Subtitle-Generator/requirements.txt
flask
google-cloud-speech
google-cloud-translate==2.0.1
moviepy
ffmpeg
regex
